requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
exceptiongroup==1.1.2
fastapi==0.100.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
multidict==6.0.4
openai==0.27.8
pydantic==2.0.2
pydantic_core==2.1.2
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
yarl==1.9.2
